Create Controlling an Arduino from Unity LED
Once the circuit has been created, you’ll need to open the RYG.ino file in Arduino and compile, then upload it. 
This is just a simple script that will change the LEDs over every second and make sure that the circuit is correct.
Around this point you should note the COM port the Arduino is on (Tools/Serial Port).